requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.4
Flask-Cors>=4.0.0
passlib>=1.7
marshmallow>=3.0
Flask-Limiter>=2.0
psycopg2-binary>=2.9
gunicorn>=20
python-dotenv>=1.0.0
